refactor: modularize utility functions and component props, remove unused code
