+++ src/library/constants/constants.ts
@@ -1,7 +1,10 @@
 import type { PacificCountry, RatesOfChangeYear } from '../types'
 
-// Responsive constants
 export const DEFAULT_MOBILE_WIDTH_THRESHOLD = 768
+export const NONE_VALUE = 'none'
+export const SEARCHBAR_INITIAL_VALUE = 'Select location for coastline data'
+export const COUNTRY_DATA_URL =
+  'https://dep-public-staging.s3.us-west-2.amazonaws.com/dep_ls_coastlines/dashboard_stats/0-0-1/country_summaries.geojson'
 
 export const RATES_OF_CHANGE_YEARS: RatesOfChangeYear[] = [
   { id: '2023', value: '2023' },
@@ -30,10 +33,6 @@ export const RATES_OF_CHANGE_YEARS: RatesOfChangeYear[] = [
   { id: '2000', value: '2000' },
   { id: '1999', value: '1999' },
 ] as const
-
-export const NONE_VALUE = 'none'
-
-export const SEARCHBAR_INITIAL_VALUE = 'Select location for coastline data'
 
 export const PACIFIC_COUNTRIES_NAMES: PacificCountry[] = [
   { id: 'ASM', name: 'American Samoa' },
